feat: move SimpleAnimationScreen to anim pkg
