+++ anim/go.mod
@@ -1,18 +1,19 @@
-module github.com/rwxrob/bonzai/run
+module github.com/rwxrob/bonzai/anim
 
-go 1.23.0
+go 1.23.2
 
 require (
-	github.com/rwxrob/bonzai/futil v0.1.0
-	github.com/rwxrob/bonzai/mark v0.1.0
+	github.com/rwxrob/bonzai/run v0.3.1
+	github.com/rwxrob/bonzai/term v0.2.1
 )
 
 require (
 	github.com/rogpeppe/go-internal v1.13.1 // indirect
 	github.com/rwxrob/bonzai/ds v0.1.1 // indirect
 	github.com/rwxrob/bonzai/fn v0.1.0 // indirect
+	github.com/rwxrob/bonzai/futil v0.1.0 // indirect
+	github.com/rwxrob/bonzai/mark v0.1.0 // indirect
 	github.com/rwxrob/bonzai/scanner v0.1.0 // indirect
-	github.com/rwxrob/bonzai/term v0.2.1 // indirect
 	github.com/rwxrob/bonzai/to v0.1.0 // indirect
 	github.com/rwxrob/bonzai/uniq v0.1.0 // indirect
 	golang.org/x/crypto v0.28.0 // indirect
