+++ run/run.go
@@ -16,7 +16,6 @@ import (
 
 	"github.com/rwxrob/bonzai/futil"
 	"github.com/rwxrob/bonzai/mark"
-	"github.com/rwxrob/bonzai/term/esc"
 )
 
 var ExePath = os.Executable
@@ -320,21 +319,4 @@ func Trap(handler func(), signals ...os.Signal) {
 		<-signalChannel
 		handler()
 	}()
-}
-
-// SimpleAnimationScreen conveniently sets up an alternate screen buffer
-// clears it, turns off the cursor and traps any interrupts so that the
-// screen and cursor are restored and the program exits. This is useful
-// when making simple ASCII animations without needing a full terminal
-// animation package.
-func SimpleAnimationScreen() error {
-	Trap(func() {
-		fmt.Print(esc.Clear)     // Clear terminal
-		fmt.Print(esc.AltBufOff) // Show cursor
-		fmt.Print(esc.CursorOn)  // Show cursor
-		Exit()
-	}, syscall.SIGINT, syscall.SIGTERM)
-	fmt.Print(esc.CursorOff)
-	fmt.Print(esc.AltBufOn)
-	return nil
 }
